Read bitmap file header
A bitmap file header is a block of bytes that is at the start of the
file and is used to identify the file. A typical application reads
this block first to ensure that the file is actually a BMP file and
that it is not damaged. The first 2 bytes of the BMP file format are
the character "B" then the character "M" in ASCII encoding.