Read main surface data
The first direct draw block compression format (BC1) works on 4×4 blocks
of texels. Instead of storing 16 colors, the algorithm saves 2 reference
colors (color_0 and color_1) and 16 2-bit color indices. The color
indices are used to look up the original colors from a color table. The
color table contains 4 colors. The first two colors, color_0 and
color_1, are the minimum and maximum colors. The other two colors,
color_2 and color_3, are intermediate colors calculated with linear
interpolation.

  color_2 = 2/3*color_0 + 1/3*color_1
  color_3 = 1/3*color_0 + 2/3*color_1

The four colors are assigned 2-bit index values that will be saved in
a 4x4 lookup table.

  color_0 = 00
  color_1 = 01
  color_2 = 10
  color_3 = 11

Finally, each of the colors in the original 4x4 block are compared with
the four colors in the color table, and the index for the closest color
is stored in the 2-bit lookup table.