Fix dds reference color handling
The reference colors (as pretty much everything else too) in dds are
handled as a single unsigned 32-bit value, e.g.:

  [32-bit ref] = [16-bit color 0][16-bit color 1]

This causes problems when writing this value to disk in little endian
architectures (like in x86), the value in disk will have its high and
low parts swapped, e.g.:

  [32-bit ref] = [16-bit color 1][16-bit color 0]

To counter this the high and low parts of the 32-bit value will have
to be swapped when written to disk in little endian systems. Similar
kind of fix is needed when reading the value from disk.