Replace local copy of Google Test with a submodule
This project uses the Google Test unit testing framework to run the unit
tests. The framework was originally copied as part of the project but to
make the maintenance of this project easier it's now replaced with a git
submodule. This way we can easily update the framework with the upstream
version with a few git commands. See more about git submodules from this
link: https://git-scm.com/book/en/v2/Git-Tools-Submodules