Make padding in bitmap pixel storage dynamic
The color data in bitmap image format is packed in rows and each row is
rounded up to a multiple of 4 bytes by padding. The amount of padding in
a row is calculated using the color depth (bits per pixel) and the image
width. Due to the dynamic nature of padding (though this value is not
changed at the current version of this project) this value is no longer
stored as such but calculated dynamically when needed. This also reduces
code duplication as both the reader and the converter had implemented
this.