Add reader and writer classes for the images
The handling of the bitmap and direct draw surface image format is
somewhat difficult because of the tightly coupled file operations. In an
attempt to make things easier in the future the file operations are now
handled via separate reader and writer classes while still keeping the
actual serialization operations intact.