Simplify bitmap pixel array file operations
The bitmap pixel array file operations were needlessly complicated as
the conversions between the color table and the pixel array format was
done as part of the file operation. After this change the pixel array
is stored as such within the Bitmap making the file IO really simple.
New member functions will handle the conversions between the color
table and the pixel array.