Add missing support for 1-bit alpha in dds
DXT1 saves 2 reference colors (c0 and c1) and 16 2-bit color indices.
The color indices are used to look up the original colors from a 4 color
color table. If c0 > c1 then c2 and c3 are calculated with linear
interpolation. If c0 <= c1 then c2 is a linear blend of the first two
colors and c3 is a transparent color.