Refactor block compressor
The block compressor was originally designed to support any standard
container type by heavily using templates. This made the header file
quite messy and hard to read. Since the rest of the code base only
uses vectors the templates can be removed and the code moved to the
source file making the code much cleaner. Once the code was moved it
was refactored quite heavily to make to code more compact abd easier
to read.