Improve compression quality by using Euclidean distance
The use of absolute color values for the reference colors and in lookup
table creation yields a fast compression speed in the expense of image
quality. By breaking the colors into their basic components and using
the Euclidean distance (https://en.wikipedia.org/wiki/Euclidean_distance)
to find the closest or the furthermost colors improves the image quality
considerably in the expense of compression speed.