Add a new texel data type for direct draw surface
To make it easier to work with dds data types the raw uint32_t data type
within the struct is replaced with a texel structure that defines the
reference colors as well as the lookup table. As a bonus feature after
this change the reference colors align correctly when read from or saved
to a file without the extra swap operation.