Add a block iterator to reduce copying of data
Bitmap and direct draw surface format both store the pixels in an image
in their own way; bitmap as a continuous stream of colors starting from
the bottom left corner of the image, direct draw surface as a 4x4 blocks
of texels starting from the top left corner of the image. On top of this
color depth varies between the format. All this means that the data has
to be copied one or more times between the formats and the more times
the data is copied over the slower the conversion speed. To reduce
copying between the formats a new block iterator adaptor is added that
allows the stream of colors to be viewed as 4x4 blocks of texels and
vice versa.