Add explicit data type for high color
To improve code readability an explicit data type is added for high
color type. The underlying type is still uint16_t but with an alias
the code readability is improved a lot. This should also make it
easier to add a proper data type (similar to true color) for high
color in the future.