Add a Docker development image
To separate the build process from the build system a support for Docker
based development is added. A new `Dockerfile` text document is added
that contains all the commands to assemble a development image. The
image includes all the components needed to build, test and document the
project.

See README.md for more information about how to build with Docker.