Refactor the image converter
The image converter has been pretty static since it was first written.
To bring it up-to-date with the rest of the code base the whole
implementation is refactored and simplified. This also revealed that the
main thing the image converter does is creating headers for the image
types meaning it probably should be refactored more or even rewritten
completely.