Extract the core block codec into its own class
The first direct draw block compression format (BC1) works on 4×4 blocks
of texels; to compress an image is to compress each block of the image.
To reflect this conceptual difference in abstraction levels in the code
the core block codec is extracted into its own module from the full
image compressor.