Move Texel out of DirectDrawSurface
The scope of the current block compressor is a bit too wide; it knows
about the different file formats (bmp and dds) and how to compress and
decompress data between them. As a first step towards creating a codec
that implements the compression algorithm without any knowledge of the
file formats the Texel data structure is moved out of the
DirectDrawSurface class.