Separate decompression and upscaling
To avoid copying the decompressed data the upscaling of the 16-bit color
pixels into 24-bit was added as part of the decompression algorithm. As
upscaling is really not part of decompression process a new transform
iterator adaptor is added that injects the color upscaling into the
decompression process while also hiding the it from the decompressor.