Fix compilation issues with Clang
Well-written code should be compiler-independent; it produces working
software regardless of the compiler used build it. Different compilers
usually produce different erros and warnings; the use of more than one
compiler to build the software should therefore lead to a wider coverage
of errors and warnings and thus also to better code quality.