Open input file only once
Because the input file is processed based on the contents of the file,
not the file name, there is no way to know in advance whether the reader
used to process the file is correct for the file. The original way to
handle this was to reopen the file and process it with a second reader
if the first reader was not correct. A more efficient way is to reuse an
already open file (as long as it is first rewound) and process it with a
second reader. (This also forces the somewhat strange use of rvalue
references to be removed in file streams.)