Combine separate compressor and decompressor into a codec
Having compressor and decompressor codes in separate classes based on
input types is a perfectly valid first implementation. However, it
violates the Single-responsibility principle which states that code with
the same reason for a change should be entirely encapsulated by a class,
module or function. A change in the compression algorithm must likely be
met with a similar change in the decompression algorithm.

  https://en.wikipedia.org/wiki/Single-responsibility_principle